Vectorize PST feature index generation on accumulator refresh (AVX512) (#792)
VSTC
Elo   | 2.70 +- 1.96 (95%)
SPRT  | 4.0+0.04s Threads=1 Hash=16MB
LLR   | 2.97 (-2.25, 2.89) [0.00, 3.00]
Games | N: 31456 W: 8063 L: 7819 D: 15574
Penta | [122, 3445, 8374, 3641, 146]
https://recklesschess.space/test/12369/

Bench: 2795637

+++ src/nnue.rs
@@ -66,7 +66,7 @@ const FT_SHIFT: i32 = 9;
 const DEQUANT_MULTIPLIER: f32 = (1 << FT_SHIFT) as f32 / (FT_QUANT * FT_QUANT * L1_QUANT) as f32;
 
 #[rustfmt::skip]
-const INPUT_BUCKETS_LAYOUT: [usize; 64] = [
+const INPUT_BUCKETS_LAYOUT: [u8; 64] = [
     0, 1, 2, 3, 3, 2, 1, 0,
     4, 5, 6, 7, 7, 6, 5, 4,
     8, 8, 8, 8, 8, 8, 8, 8,
